start refactoring Plot, Subplot, etc. components
