+++ src/Commons.jl
@@ -3,6 +3,8 @@ module Commons
 export AVec, AMat, KW, AKW, TicksArgs
 export PLOTS_SEED, PX_PER_INCH, DPI, MM_PER_INCH, MM_PER_PX
 export _haligns, _valigns
+# Functions
+export get_subplot, coords, ispolar, _update_axis, expand_extrema!, series_list
 
 const AVec = AbstractVector
 const AMat = AbstractMatrix
@@ -18,4 +20,10 @@ const MM_PER_PX   = MM_PER_INCH / PX_PER_INCH
 const _haligns = :hcenter, :left, :right
 const _valigns = :vcenter, :top, :bottom
 
+function get_subplot end
+function series_list end
+function coords end
+function ispolar end
+function _update_axis end
+function expand_extrema! end
 end
