+++ src/args.jl
@@ -122,10 +122,6 @@ like_histogram(seriestype::Symbol) = seriestype in _histogram_like
 like_line(seriestype::Symbol)      = seriestype in _line_like
 like_surface(seriestype::Symbol)   = RecipesPipeline.is_surface(seriestype)
 
-RecipesPipeline.is3d(series::Series) = RecipesPipeline.is3d(series.plotattributes)
-RecipesPipeline.is3d(sp::Subplot) = string(sp.attr[:projection]) == "3d"
-ispolar(sp::Subplot) = string(sp.attr[:projection]) == "polar"
-ispolar(series::Series) = ispolar(series.plotattributes[:subplot])
 
 # ------------------------------------------------------------
 
@@ -1517,8 +1513,6 @@ function preprocess_attributes!(plotattributes::AKW)
     end
     nothing
 end
-RecipesPipeline.preprocess_attributes!(plt::Plot, plotattributes::AKW) =
-    Plots.preprocess_attributes!(plotattributes)
 
 # -----------------------------------------------------------------------------
 
@@ -1654,7 +1648,6 @@ end
 
 # 1-row matrices will give an element
 # multi-row matrices will give a column
-# InputWrapper just gives the contents
 # anything else is returned as-is
 function slice_arg(v::AMat, idx::Int)
     isempty(v) && return v
@@ -1662,7 +1655,6 @@ function slice_arg(v::AMat, idx::Int)
     m, n = axes(v)
     size(v, 1) == 1 ? v[first(m), n[c]] : v[:, n[c]]
 end
-slice_arg(wrapper::InputWrapper, idx) = wrapper.obj
 slice_arg(v::NTuple{2,AMat}, idx::Int) = slice_arg(v[1], idx), slice_arg(v[2], idx)
 slice_arg(v, idx) = v
 
@@ -1740,49 +1732,6 @@ const _match_map2 = Dict(
     :guidefontfamily => :fontfamily_subplot,
 )
 
-# properly retrieve from plt.attr, passing `:match` to the correct key
-Base.getindex(plt::Plot, k::Symbol) =
-    if (v = plt.attr[k]) === :match
-        plt[_match_map[k]]
-    else
-        v
-    end
-
-# properly retrieve from sp.attr, passing `:match` to the correct key
-Base.getindex(sp::Subplot, k::Symbol) =
-    if (v = sp.attr[k]) === :match
-        if haskey(_match_map2, k)
-            sp.plt[_match_map2[k]]
-        else
-            sp[_match_map[k]]
-        end
-    else
-        v
-    end
-
-# properly retrieve from axis.attr, passing `:match` to the correct key
-Base.getindex(axis::Axis, k::Symbol) =
-    if (v = axis.plotattributes[k]) === :match
-        if haskey(_match_map2, k)
-            axis.sps[1][_match_map2[k]]
-        else
-            axis[_match_map[k]]
-        end
-    else
-        v
-    end
-
-Base.getindex(series::Series, k::Symbol) = series.plotattributes[k]
-
-Base.setindex!(plt::Plot, v, k::Symbol)      = (plt.attr[k] = v)
-Base.setindex!(sp::Subplot, v, k::Symbol)    = (sp.attr[k] = v)
-Base.setindex!(axis::Axis, v, k::Symbol)     = (axis.plotattributes[k] = v)
-Base.setindex!(series::Series, v, k::Symbol) = (series.plotattributes[k] = v)
-
-Base.get(plt::Plot, k::Symbol, v)      = get(plt.attr, k, v)
-Base.get(sp::Subplot, k::Symbol, v)    = get(sp.attr, k, v)
-Base.get(axis::Axis, k::Symbol, v)     = get(axis.plotattributes, k, v)
-Base.get(series::Series, k::Symbol, v) = get(series.plotattributes, k, v)
 
 # -----------------------------------------------------------------------------
 
@@ -1796,170 +1745,10 @@ function fg_color(plotattributes::AKW)
     end
 end
 
-# update attr from an input dictionary
-function _update_plot_args(plt::Plot, plotattributes_in::AKW)
-    for (k, v) in _plot_defaults
-        slice_arg!(plotattributes_in, plt.attr, k, 1, true)
-    end
-
-    # handle colors
-    plt[:background_color] = plot_color(plt.attr[:background_color])
-    plt[:foreground_color] = fg_color(plt.attr)
-    color_or_nothing!(plt.attr, :background_color_outside)
-end
 
 # -----------------------------------------------------------------------------
 
-function _update_subplot_periphery(sp::Subplot, anns::AVec)
-    # extend annotations, and ensure we always have a (x,y,PlotText) tuple
-    newanns = []
-    for ann in vcat(anns, sp[:annotations])
-        append!(newanns, process_annotation(sp, ann))
-    end
-    sp.attr[:annotations] = newanns
 
-    # handle legend/colorbar
-    sp.attr[:legend_position] = convertLegendValue(sp.attr[:legend_position])
-    sp.attr[:colorbar] = convertLegendValue(sp.attr[:colorbar])
-    if sp.attr[:colorbar] === :legend
-        sp.attr[:colorbar] = sp.attr[:legend_position]
-    end
-    nothing
-end
-
-function _update_subplot_colors(sp::Subplot)
-    # background colors
-    color_or_nothing!(sp.attr, :background_color_subplot)
-    sp.attr[:color_palette] = get_color_palette(sp.attr[:color_palette], 30)
-    color_or_nothing!(sp.attr, :legend_background_color)
-    color_or_nothing!(sp.attr, :background_color_inside)
-
-    # foreground colors
-    color_or_nothing!(sp.attr, :foreground_color_subplot)
-    color_or_nothing!(sp.attr, :legend_foreground_color)
-    color_or_nothing!(sp.attr, :foreground_color_title)
-    nothing
-end
-
-_update_margins(sp::Subplot) =
-    for sym in (:margin, :left_margin, :top_margin, :right_margin, :bottom_margin)
-        if (margin = get(sp.attr, sym, nothing)) isa Tuple
-            # transform e.g. (1, :mm) => 1 * Plots.mm
-            sp.attr[sym] = margin[1] * getfield(@__MODULE__, margin[2])
-        end
-    end
-
-function _update_axis(
-    plt::Plot,
-    sp::Subplot,
-    plotattributes_in::AKW,
-    letter::Symbol,
-    subplot_index::Int,
-)
-    # get (maybe initialize) the axis
-    axis = get_axis(sp, letter)
-
-    _update_axis(axis, plotattributes_in, letter, subplot_index)
-
-    # convert a bool into auto or nothing
-    if isa(axis[:ticks], Bool)
-        axis[:ticks] = axis[:ticks] ? :auto : nothing
-    end
-
-    _update_axis_colors(axis)
-    _update_axis_links(plt, axis, letter)
-    nothing
-end
-
-function _update_axis(
-    axis::Axis,
-    plotattributes_in::AKW,
-    letter::Symbol,
-    subplot_index::Int,
-)
-    # build the KW of arguments from the letter version (i.e. xticks --> ticks)
-    kw = KW()
-    for k in _all_axis_args
-        # first get the args without the letter: `tickfont = font(10)`
-        # note: we don't pop because we want this to apply to all axes! (delete after all have finished)
-        if haskey(plotattributes_in, k)
-            kw[k] = slice_arg(plotattributes_in[k], subplot_index)
-        end
-
-        # then get those args that were passed with a leading letter: `xlabel = "X"`
-        lk = get_attr_symbol(letter, k)
-
-        if haskey(plotattributes_in, lk)
-            kw[k] = slice_arg(plotattributes_in[lk], subplot_index)
-        end
-    end
-
-    # update the axis
-    attr!(axis; kw...)
-    nothing
-end
-
-function _update_axis_colors(axis::Axis)
-    # # update the axis colors
-    color_or_nothing!(axis.plotattributes, :foreground_color_axis)
-    color_or_nothing!(axis.plotattributes, :foreground_color_border)
-    color_or_nothing!(axis.plotattributes, :foreground_color_guide)
-    color_or_nothing!(axis.plotattributes, :foreground_color_text)
-    color_or_nothing!(axis.plotattributes, :foreground_color_grid)
-    color_or_nothing!(axis.plotattributes, :foreground_color_minor_grid)
-    nothing
-end
-
-function _update_axis_links(plt::Plot, axis::Axis, letter::Symbol)
-    # handle linking here.  if we're passed a list of
-    # other subplots to link to, link them together
-    (link = axis[:link]) |> isempty && return
-    for other_sp in link
-        link_axes!(axis, get_axis(get_subplot(plt, other_sp), letter))
-    end
-    axis.plotattributes[:link] = []
-    nothing
-end
-
-# update a subplots args and axes
-function _update_subplot_args(
-    plt::Plot,
-    sp::Subplot,
-    plotattributes_in,
-    subplot_index::Int,
-    remove_pair::Bool,
-)
-    anns = RecipesPipeline.pop_kw!(sp.attr, :annotations)
-
-    # grab those args which apply to this subplot
-    for k in keys(_subplot_defaults)
-        slice_arg!(plotattributes_in, sp.attr, k, subplot_index, remove_pair)
-    end
-
-    _update_subplot_colors(sp)
-    _update_margins(sp)
-    colorbar_update_keys =
-        (:clims, :colorbar, :seriestype, :marker_z, :line_z, :fill_z, :colorbar_entry)
-    if any(haskey.(Ref(plotattributes_in), colorbar_update_keys))
-        _update_subplot_colorbars(sp)
-    end
-
-    lims_warned = false
-    for letter in (:x, :y, :z)
-        _update_axis(plt, sp, plotattributes_in, letter, subplot_index)
-        lk = get_attr_symbol(letter, :lims)
-
-        # warn against using `Range` in x,y,z lims
-        if !lims_warned &&
-           haskey(plotattributes_in, lk) &&
-           plotattributes_in[lk] isa AbstractRange
-            @warn "lims should be a Tuple, not $(typeof(plotattributes_in[lk]))."
-            lims_warned = true
-        end
-    end
-
-    _update_subplot_periphery(sp, anns)
-end
 
 # -----------------------------------------------------------------------------
 
@@ -1971,19 +1760,6 @@ has_black_border_for_default(st::Function) =
 has_black_border_for_default(st::Symbol) =
     like_histogram(st) || st in (:hexbin, :bar, :shape)
 
-# converts a symbol or string into a Colorant or ColorGradient
-# and assigns a color automatically
-get_series_color(c, sp::Subplot, n::Int, seriestype) =
-    if c === :auto
-        like_surface(seriestype) ? cgrad() : _cycle(sp[:color_palette], n)
-    elseif isa(c, Int)
-        _cycle(sp[:color_palette], c)
-    else
-        c
-    end |> plot_color
-
-get_series_color(c::AbstractArray, sp::Subplot, n::Int, seriestype) =
-    map(x -> get_series_color(x, sp, n, seriestype), c)
 
 ensure_gradient!(plotattributes::AKW, csym::Symbol, asym::Symbol) =
     if plotattributes[csym] isa ColorPalette
@@ -2006,14 +1782,6 @@ _replace_linewidth(plotattributes::AKW) =
             DEFAULT_LINEWIDTH[]
     end
 
-function _slice_series_args!(plotattributes::AKW, plt::Plot, sp::Subplot, commandIndex::Int)
-    for k in keys(_series_defaults)
-        haskey(plotattributes, k) &&
-            slice_arg!(plotattributes, plotattributes, k, commandIndex, false)
-    end
-    plotattributes
-end
-
 label_to_string(label::Bool, series_plotindex) =
     label ? label_to_string(:auto, series_plotindex) : ""
 label_to_string(label::Nothing, series_plotindex) = ""
@@ -2028,91 +1796,6 @@ label_to_string(label::Symbol, series_plotindex) =
     end
 label_to_string(label, series_plotindex) = string(label)  # Fallback to string promotion
 
-function _update_series_attributes!(plotattributes::AKW, plt::Plot, sp::Subplot)
-    pkg = plt.backend
-    globalIndex = plotattributes[:series_plotindex]
-    plotIndex = _series_index(plotattributes, sp)
-
-    aliasesAndAutopick(
-        plotattributes,
-        :linestyle,
-        _styleAliases,
-        supported_styles(pkg),
-        plotIndex,
-    )
-    aliasesAndAutopick(
-        plotattributes,
-        :markershape,
-        _markerAliases,
-        supported_markers(pkg),
-        plotIndex,
-    )
-
-    # update alphas
-    for asym in (:linealpha, :markeralpha, :fillalpha)
-        if plotattributes[asym] === nothing
-            plotattributes[asym] = plotattributes[:seriesalpha]
-        end
-    end
-    if plotattributes[:markerstrokealpha] === nothing
-        plotattributes[:markerstrokealpha] = plotattributes[:markeralpha]
-    end
-
-    # update series color
-    scolor = plotattributes[:seriescolor]
-    stype = plotattributes[:seriestype]
-    plotattributes[:seriescolor] = scolor = get_series_color(scolor, sp, plotIndex, stype)
-
-    # update other colors (`linecolor`, `markercolor`, `fillcolor`) <- for grep
-    for s in (:line, :marker, :fill)
-        csym, asym = Symbol(s, :color), Symbol(s, :alpha)
-        plotattributes[csym] = if plotattributes[csym] === :auto
-            plot_color(if has_black_border_for_default(stype) && s === :line
-                sp[:foreground_color_subplot]
-            else
-                scolor
-            end)
-        elseif plotattributes[csym] === :match
-            plot_color(scolor)
-        else
-            get_series_color(plotattributes[csym], sp, plotIndex, stype)
-        end
-    end
-
-    # update markerstrokecolor
-    plotattributes[:markerstrokecolor] = if plotattributes[:markerstrokecolor] === :match
-        plot_color(sp[:foreground_color_subplot])
-    elseif plotattributes[:markerstrokecolor] === :auto
-        get_series_color(plotattributes[:markercolor], sp, plotIndex, stype)
-    else
-        get_series_color(plotattributes[:markerstrokecolor], sp, plotIndex, stype)
-    end
-
-    # if marker_z, fill_z or line_z are set, ensure we have a gradient
-    if plotattributes[:marker_z] !== nothing
-        ensure_gradient!(plotattributes, :markercolor, :markeralpha)
-    end
-    if plotattributes[:line_z] !== nothing
-        ensure_gradient!(plotattributes, :linecolor, :linealpha)
-    end
-    if plotattributes[:fill_z] !== nothing
-        ensure_gradient!(plotattributes, :fillcolor, :fillalpha)
-    end
-
-    # scatter plots don't have a line, but must have a shape
-    if plotattributes[:seriestype] in (:scatter, :scatterbins, :scatterhist, :scatter3d)
-        plotattributes[:linewidth] = 0
-        if plotattributes[:markershape] === :none
-            plotattributes[:markershape] = :circle
-        end
-    end
-
-    # set label
-    plotattributes[:label] = label_to_string.(plotattributes[:label], globalIndex)
-
-    _replace_linewidth(plotattributes)
-    plotattributes
-end
 
 _series_index(plotattributes, sp) =
     if haskey(plotattributes, :series_index)
