+++ src/axes_utils.jl
@@ -1,116 +1,3 @@
-
-# xaxis(args...; kw...) = Axis(:x, args...; kw...)
-# yaxis(args...; kw...) = Axis(:y, args...; kw...)
-# zaxis(args...; kw...) = Axis(:z, args...; kw...)
-
-# -------------------------------------------------------------------------
-
-function Axis(sp::Subplot, letter::Symbol, args...; kw...)
-    explicit = KW(
-        :letter => letter,
-        :extrema => Extrema(),
-        :discrete_map => Dict(),   # map discrete values to discrete indices
-        :continuous_values => zeros(0),
-        :discrete_values => [],
-        :use_minor => false,
-        :show => true,  # show or hide the axis? (useful for linked subplots)
-    )
-
-    attr = DefaultsDict(explicit, _axis_defaults_byletter[letter])
-
-    # update the defaults
-    attr!(Axis([sp], attr), args...; kw...)
-end
-
-function get_axis(sp::Subplot, letter::Symbol)
-    axissym = get_attr_symbol(letter, :axis)
-    if haskey(sp.attr, axissym)
-        sp.attr[axissym]
-    else
-        sp.attr[axissym] = Axis(sp, letter)
-    end::Axis
-end
-
-function process_axis_arg!(plotattributes::AKW, arg, letter = "")
-    T = typeof(arg)
-    arg = get(_scaleAliases, arg, arg)
-    if typeof(arg) <: Font
-        plotattributes[get_attr_symbol(letter, :tickfont)] = arg
-        plotattributes[get_attr_symbol(letter, :guidefont)] = arg
-
-    elseif arg in _allScales
-        plotattributes[get_attr_symbol(letter, :scale)] = arg
-
-    elseif arg in (:flip, :invert, :inverted)
-        plotattributes[get_attr_symbol(letter, :flip)] = true
-
-    elseif T <: AbstractString
-        plotattributes[get_attr_symbol(letter, :guide)] = arg
-
-        # xlims/ylims
-    elseif (T <: Tuple || T <: AVec) && length(arg) == 2
-        sym = typeof(arg[1]) <: Number ? :lims : :ticks
-        plotattributes[get_attr_symbol(letter, sym)] = arg
-
-        # xticks/yticks
-    elseif T <: AVec
-        plotattributes[get_attr_symbol(letter, :ticks)] = arg
-
-    elseif arg === nothing
-        plotattributes[get_attr_symbol(letter, :ticks)] = []
-
-    elseif T <: Bool || arg in _allShowaxisArgs
-        plotattributes[get_attr_symbol(letter, :showaxis)] = showaxis(arg, letter)
-
-    elseif typeof(arg) <: Number
-        plotattributes[get_attr_symbol(letter, :rotation)] = arg
-
-    elseif typeof(arg) <: Function
-        plotattributes[get_attr_symbol(letter, :formatter)] = arg
-
-    elseif !handleColors!(
-        plotattributes,
-        arg,
-        get_attr_symbol(letter, :foreground_color_axis),
-    )
-        @warn "Skipped $(letter)axis arg $arg"
-    end
-end
-
-# update an Axis object with magic args and keywords
-function attr!(axis::Axis, args...; kw...)
-    # first process args
-    plotattributes = axis.plotattributes
-    foreach(arg -> process_axis_arg!(plotattributes, arg), args)
-
-    # then preprocess keyword arguments
-    Plots.preprocess_attributes!(KW(kw))
-
-    # then override for any keywords... only those keywords that already exists in plotattributes
-    for (k, v) in kw
-        haskey(plotattributes, k) || continue
-        if k === :discrete_values
-            foreach(x -> discrete_value!(axis, x), v)  # add these discrete values to the axis
-        elseif k === :lims && isa(v, NTuple{2,TimeType})
-            plotattributes[k] = (v[1].instant.periods.value, v[2].instant.periods.value)
-        else
-            plotattributes[k] = v
-        end
-    end
-
-    # replace scale aliases
-    if haskey(_scaleAliases, plotattributes[:scale])
-        plotattributes[:scale] = _scaleAliases[plotattributes[:scale]]
-    end
-
-    axis
-end
-
-# -------------------------------------------------------------------------
-
-Base.show(io::IO, axis::Axis) = dumpdict(io, axis.plotattributes, "Axis")
-ignorenan_extrema(axis::Axis) = (ex = axis[:extrema]; (ex.emin, ex.emax))
-
 const _label_func =
     Dict{Symbol,Function}(:log10 => x -> "10^$x", :log2 => x -> "2^$x", :ln => x -> "e^$x")
 labelfunc(scale::Symbol, backend::AbstractBackend) = get(_label_func, scale, string)
@@ -208,29 +95,6 @@ function get_labels(formatter::Function, scaled_ticks, scale)
     filter!(!ismissing, fticks)
     eltype(fticks) <: Number && return get_labels(:auto, map(sf, fticks), scale)
     return fticks
-end
-
-# returns (continuous_values, discrete_values) for the ticks on this axis
-function get_ticks(sp::Subplot, axis::Axis; update = true, formatter = axis[:formatter])
-    if update || !haskey(axis.plotattributes, :optimized_ticks)
-        dvals = axis[:discrete_values]
-        ticks = _transform_ticks(axis[:ticks], axis)
-        axis.plotattributes[:optimized_ticks] =
-            if (
-                axis[:letter] === :x &&
-                ticks isa Symbol &&
-                ticks !== :none &&
-                !isempty(dvals) &&
-                ispolar(sp)
-            )
-                collect(0:(π / 4):(7π / 4)), string.(0:45:315)
-            else
-                cvals = axis[:continuous_values]
-                alims = axis_limits(sp, axis[:letter])
-                get_ticks(ticks, cvals, dvals, alims, axis[:scale], formatter)
-            end
-    end
-    axis.plotattributes[:optimized_ticks]
 end
 
 # Ticks getter functions
@@ -412,134 +276,6 @@ end
 
 # -------------------------------------------------------------------------
 
-function reset_extrema!(sp::Subplot)
-    for asym in (:x, :y, :z)
-        sp[get_attr_symbol(asym, :axis)][:extrema] = Extrema()
-    end
-    for series in sp.series_list
-        expand_extrema!(sp, series.plotattributes)
-    end
-end
-
-function expand_extrema!(ex::Extrema, v::Number)
-    ex.emin = isfinite(v) ? min(v, ex.emin) : ex.emin
-    ex.emax = isfinite(v) ? max(v, ex.emax) : ex.emax
-    ex
-end
-
-expand_extrema!(axis::Axis, v::Number) = expand_extrema!(axis[:extrema], v)
-
-# these shouldn't impact the extrema
-expand_extrema!(axis::Axis, ::Nothing) = axis[:extrema]
-expand_extrema!(axis::Axis, ::Bool) = axis[:extrema]
-
-function expand_extrema!(axis::Axis, v::Tuple{MIN,MAX}) where {MIN<:Number,MAX<:Number}
-    ex = axis[:extrema]::Extrema
-    ex.emin = isfinite(v[1]) ? min(v[1], ex.emin) : ex.emin
-    ex.emax = isfinite(v[2]) ? max(v[2], ex.emax) : ex.emax
-    ex
-end
-function expand_extrema!(axis::Axis, v::AVec{N}) where {N<:Number}
-    ex = axis[:extrema]::Extrema
-    foreach(vi -> expand_extrema!(ex, vi), v)
-    ex
-end
-
-function expand_extrema!(sp::Subplot, plotattributes::AKW)
-    vert = isvertical(plotattributes)
-
-    # first expand for the data
-    for letter in (:x, :y, :z)
-        data = plotattributes[if vert
-            letter
-        else
-            letter === :x ? :y : letter === :y ? :x : :z
-        end]
-        if (
-            letter !== :z &&
-            plotattributes[:seriestype] === :straightline &&
-            any(series[:seriestype] !== :straightline for series in series_list(sp)) &&
-            length(data) > 1 &&
-            data[1] != data[2]
-        )
-            data = [NaN]
-        end
-        axis = sp[get_attr_symbol(letter, :axis)]
-
-        if isa(data, Volume)
-            expand_extrema!(sp[:xaxis], data.x_extents)
-            expand_extrema!(sp[:yaxis], data.y_extents)
-            expand_extrema!(sp[:zaxis], data.z_extents)
-        elseif eltype(data) <: Number ||
-               (isa(data, Surface) && all(di -> isa(di, Number), data.surf))
-            if !(eltype(data) <: Number)
-                # huh... must have been a mis-typed surface? lets swap it out
-                data = plotattributes[letter] = Surface(Matrix{Float64}(data.surf))
-            end
-            expand_extrema!(axis, data)
-        elseif data !== nothing
-            # TODO: need more here... gotta track the discrete reference value
-            #       as well as any coord offset (think of boxplot shape coords... they all
-            #       correspond to the same x-value)
-            plotattributes[letter],
-            plotattributes[get_attr_symbol(letter, :(_discrete_indices))] =
-                discrete_value!(axis, data)
-            expand_extrema!(axis, plotattributes[letter])
-        end
-    end
-
-    # # expand for fillrange/bar_width
-    # fillaxis, baraxis = sp.attr[:yaxis], sp.attr[:xaxis]
-    # if isvertical(plotattributes)
-    #     fillaxis, baraxis = baraxis, fillaxis
-    # end
-
-    # expand for fillrange
-    fr = plotattributes[:fillrange]
-    if fr === nothing && plotattributes[:seriestype] === :bar
-        fr = 0.0
-    end
-    if fr !== nothing && !RecipesPipeline.is3d(plotattributes)
-        axis = sp.attr[vert ? :yaxis : :xaxis]
-        if typeof(fr) <: Tuple
-            foreach(x -> expand_extrema!(axis, x), fr)
-        else
-            expand_extrema!(axis, fr)
-        end
-    end
-
-    # expand for bar_width
-    if plotattributes[:seriestype] === :bar
-        dsym = vert ? :x : :y
-        data = plotattributes[dsym]
-
-        if (bw = plotattributes[:bar_width]) === nothing
-            pos = filter(>(0), diff(sort(data)))
-            plotattributes[:bar_width] = bw = _bar_width * ignorenan_minimum(pos)
-        end
-        axis = sp.attr[get_attr_symbol(dsym, :axis)]
-        expand_extrema!(axis, ignorenan_maximum(data) + 0.5maximum(bw))
-        expand_extrema!(axis, ignorenan_minimum(data) - 0.5minimum(bw))
-    end
-
-    # expand for heatmaps
-    if plotattributes[:seriestype] === :heatmap
-        for letter in (:x, :y)
-            data = plotattributes[letter]
-            axis = sp[get_attr_symbol(letter, :axis)]
-            scale = get(plotattributes, get_attr_symbol(letter, :scale), :identity)
-            expand_extrema!(axis, heatmap_edges(data, scale))
-        end
-    end
-end
-
-function expand_extrema!(sp::Subplot, xmin, xmax, ymin, ymax)
-    expand_extrema!(sp[:xaxis], (xmin, xmax))
-    expand_extrema!(sp[:yaxis], (ymin, ymax))
-end
-
-# -------------------------------------------------------------------------
-
 function scale_lims(from, to, factor)
     mid, span = (from + to) / 2, (to - from) / 2
     mid .+ (-span, span) .* factor
@@ -559,7 +295,7 @@ end
     scale_lims!([plt], [letter], factor)
 
 Scale the limits of the axis specified by `letter` (one of `:x`, `:y`, `:z`) by the
-given `factor` around the limits' middle point. 
+given `factor` around the limits' middle point.
 If `letter` is omitted, all axes are affected.
 """
 function scale_lims!(sp::Subplot, letter, factor)
